requirements.txt
# LiveKit Agents SDK with plugin support
livekit-agents[openai,silero,deepgram,cartesia]~=0.9.0

# Additional plugins for TTS voice options
livekit-plugins-elevenlabs~=0.7.0

# Twilio for outbound calling
twilio~=9.0.0

# Environment variable management
python-dotenv~=1.0.0

# HTTP client for API calls
httpx~=0.27.0

# Database for call logging (optional)
sqlalchemy~=2.0.0
aiosqlite~=0.20.0

# Utilities
pydantic~=2.0.0
